+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Features
+
+- Add 6 new themes: Sage, Sky, and Sand (light) and Obsidian, Ocean, and Crimson (dark)
+
 ## [1.42.0]
 
 ### Features